+++ CHANGELOG.md
@@ -7,8 +7,11 @@ Versioning practices: [Semantic Versioning](https://semver.org/spec/v2.0.0.html)
 ### Fixed
 ### Added
 ### Removed
- - Remove 'stats-collect start --cpus' option.
 ### Changed
+
+## [1.0.51] - 2025-03-25
+### Removed
+ - Remove 'stats-collect start --cpus' option.
 
 ## [1.0.50] - 2025-03-14
 ### Fixed